+++ cnccode/controller_classes/CTCNC.inc.php
@@ -406,7 +406,22 @@ class CTCNC extends Controller
 
 
         if ($this->hasPermissions(SALES_PERMISSION)) {
-            $menu->addSection("Sales", 'fa-tag', $this->getDefaultSalesMenu());
+            $menu->addSection("Sales", 'fa-tag', $this->getDefaultSalesMenu());            
+            $this->addConditionalMenu(
+                $menu,
+                'fa-tag',
+                "Sales",
+                $this->dbeUser->getValue(DBEUser::streamOneLicenseManagement) == 1,
+                313,
+                "StreamOne Licenses",
+                "CustomerLicenses.php?action=searchCustomers"
+            );
+            
+            // [
+            //     "id"    => 312,
+            //     "label" => "StreamOne Licenses",
+            //     "href"  => "CustomerLicenses.php?action=searchCustomers",
+            // ],
         }
 
         $this->addConditionalMenu(
@@ -798,6 +813,12 @@ class CTCNC extends Controller
                 "label" => "Quote Templates",
                 "href"  => "QuoteTemplates.php",
             ],
+            
+            // [
+            //     "id"    => 313,
+            //     "label" => "TechData Orders",
+            //     "href"  => "CustomerLicenses.php?action=searchOrders",
+            // ],
         ];
     }
 
@@ -1201,4 +1222,5 @@ class CTCNC extends Controller
         }
         echo $js_code;
     }
+    
 }